Denester fix: Only write span begin and end times if they differ from the parent
Stops us from writing unnecessary times into spans.

This turned out to be a bit nasty, because timed elements return a computed time that can depend on their _children_. So we had to work out if the parents are specifying the end time or if it is this element doing it, and then deal with it on that basis.

Some tests updated too.