Make pushed time calculation less brittle
and pass pep8/flake8.

The previous iteration relied on there being an equal number of begin and end times pushed to the dataset and had no plan for what to do if the calling code had not done that properly. It shouldn't have been a problem but was making me uncomfortable for later.

Instead, push pairs of begin and end times as named tuples, and be less brittle about the key used to get the list from the dataset by defining it globally.

There's probably a way to refactor this even more, but this is an improvement over the first iteration, at least.