Tidy removed divs better, merge metadata
When removing divs after unnesting, leaves the `body` element in a tidy state with no orphan `div` elements in `orderedContent()`.

Treats `div`s with different metadata as being different, and merges metadata when denesting. This may not always be the best strategy, but it at least preserves the input data somehow.