Denester: correct timing type for the document timebase
Reverts `ee5845b`'s temporary workarounds to `testing/bdd/templates/nested_elements_hardcoded.xml` and `testing/bdd/templates/nested_spans_hardcoded.xml`

The reason they failed is because Denester always created a `FullClockTimingType` for any times it modified, regardless of the document's `timeBase`. That worked for `media` but not for `clock`, for which the specification requires that the more constrained `LimitedClockTimingType` is used.

Fix Denester to create the appropriate type of time depending on the document's timebase, and fix the unit tests to provide the document object in the dataset so that it can do this.

Ensures that in every test where conversion to EBU-TT-D occurs after denesting, the denested document is validated, otherwise a SemanticValidationError is (correctly) raised because some necessary computations must be done before the conversion.