Denester: make the right kind of percentage font size type
Super subtle issue that caused downstream problems. Denester was making a style to set font size, and when it made a percentage one, it was using `ebuttdt.percentageFontSizeType`; this was persisting as the computed font size after validation, however when using the computed font size downstream, for example in the `EBUTT3EBUTTDConverter`, it is important to have access to the `.vertical()` method of the derived class `PercentageFontSizeType`: not having that caused test failures. Therefore, make the right type of percentage font size type, with a capital P, and the tests pass.