ISD handling, regions with backgrounds
Previously, we processed each separate `p` in an ISD distinctly as a separate ISD, which was wrong. Now, gather all the elements in each ISD together and process as a group. Also tidies up the time handling.

When an ISD has a region with an opaque background colour and showBackground="always", there can _never_ be an empty ISD, because the background always needs to be painted.